fix(ui): restructure attack paths routing to single page
Move the main attack paths page from /attack-paths/query-builder to
/attack-paths directly. Extract page component to attack-paths-page.tsx,
relocate layout from (workflow) to attack-paths root, and update wizard
state navigation paths.
