+++ ui/app/(prowler)/attack-paths/layout.tsx
@@ -1,19 +1,18 @@
 import { Navbar } from "@/components/ui/nav-bar/navbar";
 
 /**
- * Workflow layout for Attack Paths
+ * Layout for Attack Paths
  * Displays content with navbar
  */
-export default function AttackPathsWorkflowLayout({
+export default function AttackPathsLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <>
-      <Navbar title="Attack Paths Analysis" icon="" />
+      <Navbar title="Attack Paths" icon="" />
       <div className="px-6 py-4 sm:px-8 xl:px-10">
-        {/* Content */}
         <div>{children}</div>
       </div>
     </>
