+++ ui/app/(prowler)/attack-paths/page.tsx
@@ -1,9 +1,1 @@
-import { redirect } from "next/navigation";
-
-/**
- * Landing page for Attack Paths feature
- * Redirects to the integrated attack path analysis view
- */
-export default function AttackPathsPage() {
-  redirect("/attack-paths/query-builder");
-}
+export { default } from "./(workflow)/query-builder/attack-paths-page";
